Complete EA
Fixed the issues with the mutation function and finished the EA. Also
added the weights from the individuals to the ANN and verified that it
was working properly. It seems like the EA isn’t learning as well as it
should. Further testing will be done to verify the EA.